pyproject.toml
[tool.poetry]
name = "seo-ai"
version = "0.1.0"
description = ""
authors = ["Alex Goliansky <a_goliansky@viyar.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.36"
openai = "^1.57.4"
pydantic-settings = "^2.7.0"
fastapi = {version = "0.115.5", extras = ["standard"]}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
